requirements.txt
bcrypt~=3.1.7
PyJWT~=1.4.2
Flask~=1.1.2
pymongo~=3.11.0
python-dotenv~=0.14.0
Django~=3.1
flask_pymongo
dnspython
djangorestframework==3.11.1
beautifulsoup4
google